AzQtApplication Window
