+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Window/ShaderManagementConsoleWindow.h
@@ -15,10 +15,11 @@
 #include <Atom/RPI.Edit/Shader/ShaderVariantListSourceData.h>
 
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
-#include <AzQtComponents/Components/DockMainWindow.h>
+//#include <AzQtComponents/Components/DockMainWindow.h>
 #include <AzQtComponents/Components/FancyDocking.h>
 #include <AzQtComponents/Components/StyledDockWidget.h>
 #include <AzQtComponents/Components/Widgets/TabWidget.h>
+#include <AzQtComponents/Application/Window/AzQtApplicationWindow.h>
 
 #include <Window/ShaderManagementConsoleBrowserWidget.h>
 #include <Window/ToolBar/ShaderManagementConsoleToolBar.h>
@@ -41,7 +42,7 @@ namespace ShaderManagementConsole
      * its panels, managing selection of assets, and performing high-level actions like saving. It contains...
      */
     class ShaderManagementConsoleWindow
-        : public AzQtComponents::DockMainWindow
+        : public AzQtComponents::AzQtApplicationWindow
         , private ShaderManagementConsoleDocumentNotificationBus::Handler
     {
         Q_OBJECT
@@ -59,17 +60,15 @@ namespace ShaderManagementConsole
         void OnDocumentUndoStateChanged(const AZ::Uuid& documentId) override;
         void OnDocumentSaved(const AZ::Uuid& documentId) override;
 
-        void SetupMenu();
+        void SetupMenu() override;
+        void SetupTabs() override;
 
-        void SetupTabs();
         void AddTabForDocumentId(const AZ::Uuid& documentId);
         void RemoveTabForDocumentId(const AZ::Uuid& documentId);
         void UpdateTabForDocumentId(const AZ::Uuid& documentId);
         AZ::Uuid GetDocumentIdFromTab(const int tabIndex) const;
 
-        void OpenTabContextMenu();
-        void SelectPreviousTab();
-        void SelectNextTab();
+        void OpenTabContextMenu() override;
 
         void SelectDocumentForTab(const int tabIndex);
         void CloseDocumentForTab(const int tabIndex);
@@ -79,42 +78,11 @@ namespace ShaderManagementConsole
 
         void CreateDocumentContent(const AZ::Uuid& documentId, QStandardItemModel* model);
 
-        AzQtComponents::FancyDocking* m_advancedDockManager = nullptr;
-        QMenuBar* m_menuBar = nullptr;
-        QWidget* m_centralWidget = nullptr;
-        AzQtComponents::TabWidget* m_tabWidget = nullptr;
         ShaderManagementConsoleBrowserWidget* m_assetBrowser = nullptr;
         ShaderManagementConsoleToolBar* m_toolBar = nullptr;
         AzToolsFramework::CScriptTermDialog* m_pythonTerminal = nullptr;
 
         AzQtComponents::StyledDockWidget* m_assetBrowserDockWidget = nullptr;
         AzQtComponents::StyledDockWidget* m_pythonTerminalDockWidget = nullptr;
-
-        QMenu* m_menuFile = {};
-        QMenu* m_menuNew = {};
-        QAction* m_actionOpen = {};
-        QAction* m_actionOpenRecent = {};
-        QAction* m_actionClose = {};
-        QAction* m_actionCloseAll = {};
-        QAction* m_actionCloseOthers = {};
-        QAction* m_actionSave = {};
-        QAction* m_actionSaveAsCopy = {};
-        QAction* m_actionSaveAll = {};
-        QAction* m_actionExit = {};
-
-        QMenu* m_menuEdit = {};
-        QAction* m_actionUndo = {};
-        QAction* m_actionRedo = {};
-        QAction* m_actionPreferences = {};
-
-        QMenu* m_menuView = {};
-        QAction* m_actionAssetBrowser = {};
-        QAction* m_actionPythonTerminal = {};
-        QAction* m_actionNextTab = {};
-        QAction* m_actionPreviousTab = {};
-
-        QMenu* m_menuHelp = {};
-        QAction* m_actionHelp = {};
-        QAction* m_actionAbout = {};
     };
 } // namespace ShaderManagementConsole
